Creacion capa de servicio, capa de servicio implementada para Users,Patients y Doctors

+++ MedicalAppoiments.Domain/Entities/medical/Specialties.cs
@@ -9,7 +9,7 @@ namespace MedicalAppoiments.Domain.Entities.medical
     public class Specialties : BaseEntity
     {
         [Key]
-        public int SpecialtyID { get; set; }
+        public Int16 SpecialtyID { get; set; }
         public string? SpecialtyName { get; set; }
 
         public bool IsActive { get; set; }
